clean up unused files and organize files
