--- src/modules/users/types/user.types.ts
@@ -1,5 +1,0 @@
-const UserTypes = {
-  UserRepository: Symbol.for('UserRepository')
-};
-
-export default UserTypes;
