+++ src/modules/users/useCases/signup.interface.ts
@@ -1,6 +1,6 @@
 import { Result } from '../../../infra/models/result';
+import { User } from '../domain/entities/user';
 import { SignupDto } from '../dtos/signup.dto';
-import { User } from '../models/user';
 
 export interface ISignup {
     handle(signupDto: SignupDto): Promise<Result<User>>;
